+++ packages/domain-event/src/test/__test-utils__/mocks/event.repository.mock.ts
@@ -1,14 +1,7 @@
-import { jest } from '@jest/globals';
+import { createMock } from '@volontariapp/testing';
 import type { IEventRepository } from '../../../repositories/interfaces/event.repository.js';
 
-export const createEventRepositoryMock = (): jest.Mocked<IEventRepository> =>
-  ({
-    findAll: jest.fn(),
-    findById: jest.fn(),
-    create: jest.fn(),
-    createWithEventCreated: jest.fn(),
-    update: jest.fn(),
-    delete: jest.fn(),
-    deleteWithEventDeleted: jest.fn(),
-    search: jest.fn(),
-  }) as unknown as jest.Mocked<IEventRepository>;
+export const createEventRepositoryMock = () => {
+  const mock = createMock<IEventRepository>();
+  return mock;
+};