+++ packages/database/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.7.0
+
 ## 3.3.0
 
 ### Minor Changes